2022-10-16 1.预警分-归因简报wy 开发 2.预警分-归因详情 维度风险等级为NULL 问题修复

+++ 需求开发/预警分-归因详情/预警分-归因详情历史ddl.sql
@@ -1,4 +1,4 @@
--- 预警分-归因详情历史 RMP_WARNING_SCORE_DETAIL_HIS --
+-- ?????-??????????? RMP_WARNING_SCORE_DETAIL_HIS --
 drop table if exists pth_rmp.RMP_WARNING_SCORE_DETAIL_HIS;
 create table pth_rmp.RMP_WARNING_SCORE_DETAIL_HIS
 (
@@ -12,14 +12,14 @@ create table pth_rmp.RMP_WARNING_SCORE_DETAIL_HIS
     type string,
     sub_model_name string,
     idx_name string,
-    idx_value double,
+    idx_value float,
     idx_unit string,
-    idx_score double,
-    contribution_ratio double,
+    idx_score float,
+    contribution_ratio float,
     contribution_cnt bigint,
     factor_evaluate int,
-    median  double,
-    last_idx_value double,
+    median  float,
+    last_idx_value float,
     idx_cal_explain string,
     idx_explain string,
 	delete_flag	int,
